requirements.txt
openai == 2.33.0
fastapi == 0.136.1
uvicorn == 0.46.0
python-dotenv == 1.2.2